Moves device to it's own file
